feat: add description for VIP-242
