+++ simulations/vip-242/simulations.ts
@@ -4,7 +4,7 @@ import { ethers } from "hardhat";
 
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import { COMPTROLLER, caps, vip238 } from "../../vips/vip-238";
+import { COMPTROLLER, caps, vip242 } from "../../vips/vip-242";
 import COMPTROLLER_ABI from "./abi/Comptroller.json";
 import SETTER_FACET_ABI from "./abi/SetterFacet.json";
 
@@ -15,7 +15,7 @@ forking(35215431, () => {
     comptroller = new ethers.Contract(COMPTROLLER, COMPTROLLER_ABI, ethers.provider);
   });
 
-  testVip("VIP-238", vip238(), {
+  testVip("VIP-242", vip242(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [SETTER_FACET_ABI], ["NewBorrowCap"], [13]);
       await expectEvents(txResponse, [SETTER_FACET_ABI], ["NewSupplyCap"], [11]);
